UI changes like amcl

+++ src/app/Components/logged-in/med-review/med-review.component.ts
@@ -1,7 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
+import { NgxUiLoaderService } from 'ngx-ui-loader';
+import { FormGroup, FormControl } from '@angular/forms';
+import { MatDialog, MatTableDataSource, MatSort, MatPaginator } from '@angular/material';
 
 import { PatientService } from 'src/app/Services/patient.service';
+import { DeleteConfirmationComponent } from './delete-confirmation/delete-confirmation.component';
 
 @Component({
   selector: 'app-med-review',
@@ -9,32 +13,72 @@ import { PatientService } from 'src/app/Services/patient.service';
   styleUrls: ['./med-review.component.css']
 })
 export class MedReviewComponent implements OnInit {
-
-  patientList = [];
+  filterForm: FormGroup;
+  patientList: MatTableDataSource<any>;
   displayedColumns: string[] = ['name'];
 
-  constructor(private patientService: PatientService, private router: Router) { }
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+  @ViewChild(MatSort) sort: MatSort;
+
+  constructor(private patientService: PatientService, private router: Router, public dialog: MatDialog, private ngxLoader: NgxUiLoaderService) {
+    this.filterForm = new FormGroup({
+      dobFrom: new FormControl(''),
+      dobTo: new FormControl(''),
+      search: new FormControl(''),
+    });
+  }
 
   ngOnInit() {
 
-    this.patientService.getPatientList().subscribe(
+    this.ngxLoader.start();
+
+    this.patientService.getPatientList(this.filterForm.value).subscribe(
       res => {
-        this.patientList = res;
+        this.patientList = new MatTableDataSource(res)
+        this.patientList.paginator = this.paginator
+        this.patientList.sort = this.sort
+        this.ngxLoader.stop();
+
       }
     )
   }
 
+  filters() {
+
+    this.ngxLoader.startBackground();
+
+    this.patientService.getPatientList(this.filterForm.value).subscribe(
+      res => {
+        this.patientList = new MatTableDataSource(res);
+        this.patientList.paginator = this.paginator;
+        this.patientList.sort = this.sort;
+        this.ngxLoader.stopBackground();
+      }
+    );
+  }
+
   patientDetail(id) {
-    this.router.navigate([`/review/${id}`]);
+    this.router.navigate([`/reports/${id}`]);
   };
 
   onDelete(id) {
-    this.patientService.deletePatient(id).subscribe(
-      res => {
-        // this.router.navigate[this.router.url]
-        this.ngOnInit();
+
+    const dialogRef = this.dialog.open(DeleteConfirmationComponent, {
+      width: '450px',
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+
+        this.patientService.deletePatient(id).subscribe(
+          res => {
+            this.ngOnInit();
+          }
+        )
+
       }
-    )
+    });
+
   };
 
 }
